requirements.txt
discord.py>=2.0.0
requests>=2.28.0
python-dotenv>=0.20.0
discord-ui>=5.0.0
